feat: add voting power component
